requirements-dev.txt
black
flake8
tox
python-dotenv  # optional to use .env file instead of environment variables
